Tesseract
 glows above 3 floor spiral staircase rooftop 'return' rune stone
This special item encapsulates the dignity of a 4-dimensional object by glowing magnificently in a green that matches the 'Return' runestone, which bears a related symbol: a square within another square, which is also the chinese character for 'return'. In Old MC style it hovers above the tower staircase that adjoins the Route No.7 'Sarumanti' build.