Proper gate on RN04
Hard to reconcile the flatland of RN04 with the 'Top' of RN07. The broken shell of RN06 led to the 'worm' of free squiggling. Lights are out. To quote Perec's poem (Of DSD) : '...looking back, I turn my house sacred, this is how the water melts the bottom of the air open first, in the silence of light, a blow will happen...' 
I'm thinking of moving on from github uploads now. Find me elsewhere.
-flowingpoint